Fix warnings

+++ cocos/scripting/lua-bindings/manual/cocos2d/lua_cocos2dx_deprecated.cpp
@@ -1541,7 +1541,7 @@ static int tolua_cocos2d_kmGLLoadMatrix00(lua_State* tolua_S)
             else
 #endif
             {
-                mat4.m[i] = tolua_tonumber(tolua_S, -1, 0);
+                mat4.m[i] = (float)tolua_tonumber(tolua_S, -1, 0);
                 lua_pop(tolua_S, 1);
             }
         }
